+++ calendar/config/packages/nelmio_cors.yaml
