+++ web/conf.d/calendar.conf
@@ -1,12 +1,12 @@
-upstream calendarservice {
-    server calendarservice:3000;
+upstream calendar {
+    server calendar:3000;
     keepalive 8;
 }
 
 server {
     listen 80;
     listen [::]:80;
-    server_name calendarservice.cearth.local;
+    server_name calendar.cearth.local;
 
     # pass the request to the node.js server with the correct headers
     # and much more can be added, see nginx config options
@@ -16,7 +16,7 @@ server {
       proxy_set_header Host $http_host;
       proxy_set_header X-NginX-Proxy true;
 
-      proxy_pass http://calendarservice/;
+      proxy_pass http://calendar/;
       proxy_redirect off;
     }
